requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
decorator==4.0.9
Django==1.9.6
django-admin-sortable2==0.6.3
django-angular==0.8.2
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.2.5
django-compressor==2.0
django-filer==1.2.0
django-filter==0.13.0
django-formtools==1.0
django-fsm==2.3.0
django-fsm-admin==1.2.3
django-ipware==1.1.5
django-mptt==0.8.4
django-parler==1.6.3
django-polymorphic==0.8.1
django-post-office==2.0.7
django-rest-auth==0.5.0
django-sekizai==0.9.0
Django-Select2==5.8.4
django-shop==0.9.1
django-treebeard==4.0.1
djangocms-admin-style==1.1.1
djangocms-bootstrap3==0.1.1
djangocms-cascade==0.8.4
djangocms-text-ckeditor==2.9.3
djangorestframework==3.1.3
djangoshop-stripe==0.1.4
easy-thumbnails==2.3
html5lib==0.9999999
ipython==4.2.0
ipython-genutils==0.1.0
jsonfield==1.0.3
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
stripe==1.33.0
traitlets==4.2.1
Unidecode==0.4.19
wheel==0.24.0
